# contracts/openapi/phase0.yaml
openapi: 3.0.3
info:
  title: OCRWebApp Phase 0 API
  version: 0.3.0
  description: |
    Phase 0 API契約。
    対象: documents作成、jobs状態参照、document参照、properties編集・エクスポート。
servers:
  - url: https://example.com
...
paths:
...
  /api/documents/{documentId}:
    get:
      tags:
        - documents
      summary: 文書メタデータ取得
      security:
        - sessionAuth: []
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: 所有者以外のアクセス
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  schemas:
...
    DocumentResponse:
      type: object
      required:
        - document_id
        - user_id
        - file_path
        - created_at
      properties:
        document_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        file_path:
          type: string
        created_at:
          type: string
          format: date-time
        latest_job:
          $ref: "#/components/schemas/LatestJobSummary"
          nullable: true
        normalized_property:
          $ref: "#/components/schemas/NormalizedProperty"
          nullable: true
    LatestJobSummary:
      type: object
      required:
        - job_id
        - status
        - updated_at
      properties:
        job_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/JobStatus"
        error_message:
          type: string
          nullable: true
        updated_at:
          type: string
          format: date-time
    NormalizedProperty:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          format: uuid
        property_name:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        price:
          type: number
          nullable: true
        rent:
          type: number
          nullable: true
        yield:
          type: number
          nullable: true
        structure:
          type: string
          nullable: true
        built_year:
          type: string
          nullable: true
        station_info:
          type: string
          nullable: true
        editable_fields:
          type: object
          nullable: true
        updated_at:
          type: string
          format: date-time
...
    JobStatus:
      type: string
      enum:
        - queued
        - processing
        - succeeded
        - failed
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
  responses:
...
    Unauthorized:
      description: 未認証
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 対象なし
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"